+++ app/Main.hs
@@ -1,7 +1,4 @@
 {-# LANGUAGE DataKinds #-}
-{-# LANGUAGE DeriveGeneric #-}
-{-# LANGUAGE TypeOperators #-}
-{-# LANGUAGE TypeApplications #-}
 
 import Data.Aeson
 import GHC.Generics
@@ -12,6 +9,7 @@ import System.IO
 import Control.Monad.IO.Class (MonadIO(liftIO))
 import Lib (secretKey,putEdKey,signWithEdKey, Id (Id), Label (Label),Domains(Domains), getPubKey)
 import Data.String (IsString(fromString))
+import App(app)
 
 main :: IO ()
 main = do
@@ -25,32 +23,3 @@ main = do
         defaultSettings
   runSettings settings app
 
-app :: Application
-app = serve @SignApi Proxy server
-
-type SignApi =
-  "sign" :> ReqBody '[JSON, PlainText] String :> Post '[JSON] Payload
-  :<|> "pubkey" :> Get '[JSON] Payload
-
-server :: Server SignApi
-server =
-  signTx
-  :<|> getPK
-
-signTx :: String -> Handler Payload
-signTx w = do
-  let bytes = fromString w
-  liftIO $ print bytes
-  signed <- liftIO $ signWithEdKey (Id 200) bytes True
-  return $ Payload $ show signed
-
-getPK :: Handler Payload
-getPK = do
-  pk <- liftIO $ getPubKey (Id 200) True
-  return $ Payload $ show pk
-
-newtype Payload = Payload {value :: String}
-  deriving (Eq, Show, Generic)
-
-instance ToJSON Payload
-instance FromJSON Payload
