+++ yubihsm-ed-sign.cabal
@@ -41,10 +41,17 @@ library
     exposed-modules:
       Lib
       Hex
+      App
     other-extensions: ForeignFunctionInterface
     build-depends:    base >=4.13.0.0 && <=4.18.0.0
                     , bytestring
                     , split
+                    , servant
+                    , servant-server
+                    , wai
+                    , warp
+                    , http-types
+                    , aeson
     default-language: Haskell2010
 
     extra-libraries:
